% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{ancestralMutations}
\alias{ancestralMutations}
\title{Predicted mutations on nodes}
\usage{
ancestralMutations(treeAlignMatch, similarity, ...)
}
\arguments{
\item{treeAlignMatch}{a \code{\link{treeAlignMatch}} object}

\item{similarity}{similarity threshold for tree trimming}
}
\value{
an ancestralMutation object
}
\description{
Under development
}
