% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectResult.R
\name{extractTips}
\alias{extractTips}
\alias{extractTips.fixationSites}
\alias{extractTips.multiFixationSites}
\title{Extract sitePath for a single site}
\usage{
\method{extractTips}{fixationSites}(x, site, select = 1, ...)

\method{extractTips}{multiFixationSites}(x, site, ...)
}
\arguments{
\item{x}{A \code{fixationSites} or a \code{multiFixationSites} object}

\item{site}{A site predicted to experience fixation.}

\item{select}{For a site, there theoretically might be more than one
fixation on different lineages. To extract}

\item{...}{Other arguments}
}
\value{
The name of the tips involved in the fixation
}
\description{
Retrieve the name of the tips involved in the fixation
}
