% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMutations.R
\name{findSites}
\alias{findSites}
\alias{SNPsites}
\alias{fixationSites}
\alias{fixationSites.lineagePath}
\alias{multiFixationSites}
\alias{multiFixationSites.lineagePath}
\title{Finding sites with variation}
\usage{
SNPsites(tree, minSNP = NULL)

\method{fixationSites}{lineagePath}(paths, minEffectiveSize = NULL,
  searchDepth = 1, ...)

\method{multiFixationSites}{lineagePath}(paths, minEffectiveSize = NULL,
  searchDepth = 1, ...)
}
\arguments{
\item{tree}{The return from \code{\link{addMSA}} function}

\item{minSNP}{Minimum number of amino acid variation to be a SNP}

\item{paths}{a \code{lineagePath} object returned from \code{\link{lineagePath}} function}

\item{minEffectiveSize}{A vector of two integers to specifiy minimum tree tips involved
before/after mutation. Otherwise the mutation will not be counted into
the return. If more than one number is given, the ancestral takes the first
and descendant takes the second as the minimum. If only given one number,
it's the minimum for both ancestral and descendant.}

\item{searchDepth}{The function uses heuristic search but the termination of the search
cannot be intrinsically decided. \code{searchDepth} is needed to tell
the search when to stop.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{SNPsite} returns a list of qualified SNP site

\code{fixationSites} returns a list of mutations
with names of the tips involved. The name of each list element
is the discovered mutation. A mutation has two vectors of tip names:
'from' before the fixation and 'to' after the fixation.

\code{multiFixationSites} returns sites with multiple fixations.
}
\description{
Single nucleotide polymorphism (SNP) in the whole package refers to
variation of amino acid. \code{findSNPsite} will try to find SNP in
the multiple sequence alignment. A reference sequence
and gap character may be specified to number the site. This is
irrelevant to the intended analysis but might be helpful to evaluate
the performance of \code{fixationSites}.

After finding the \code{\link{lineagePath}} of a phylogenetic tree,
\code{fixationSites} uses the result to find those sites that show
fixation on some, if not all, of the lineages. Parallel evolution is
relatively common in RNA virus. There is chance that some site be fixed
in one lineage but does not show fixation because of different
sequence context.

After finding the \code{\link{lineagePath}} of a phylogenetic tree,
\code{multiFixationSites} uses the result to find those sites that show
multiple fixations on some, if not all, of the lineages.
}
\examples{
data('zikv_tree_reduced')
data('zikv_align_reduced')
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
SNPsites(tree)
fixationSites(lineagePath(tree))
data(h3n2_tree_reduced)
data(h3n2_align_reduced)
tree <- addMSA(h3n2_tree_reduced, alignment = h3n2_align_reduced)
multiFixationSites(lineagePath(tree))
}
