% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathView.R
\name{findSites.sitePath}
\alias{findSites.sitePath}
\title{Find sites}
\usage{
\method{findSites}{sitePath}(paths, n, allseq)
}
\arguments{
\item{paths}{a \code{sitePath} object}

\item{n}{the index of the target path}

\item{allseq}{reconstructed sequences}
}
\value{
sites and tip names
}
\description{
To find site in sitePath
}
