% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixationIndels.R
\name{fixationIndels}
\alias{fixationIndels}
\alias{fixationIndels.sitesMinEntropy}
\title{Fixation indels prediction}
\usage{
fixationIndels(x, ...)

\method{fixationIndels}{sitesMinEntropy}(x, ...)
}
\arguments{
\item{x}{The return from \code{\link{sitesMinEntropy}} function.}

\item{...}{Other arguments.}
}
\value{
A \code{fixationIndels} object.
}
\description{
The fixation of insertions of deletions.
}
\examples{
data(zikv_tree_reduced)
data(zikv_align_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
fixationIndels(sitesMinEntropy(tree))
}
