% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixationSites.R
\name{fixationSites}
\alias{fixationSites}
\alias{fixationSites.lineagePath}
\title{Fixation sites prediction}
\usage{
\method{fixationSites}{lineagePath}(
  paths,
  minEffectiveSize = NULL,
  searchDepth = 1,
  method = c("compare", "insert", "delete"),
  ...
)
}
\arguments{
\item{paths}{A \code{lineagePath} object returned from
\code{\link{lineagePath}} function.}

\item{minEffectiveSize}{The minimum number of tips in a group.}

\item{searchDepth}{The function uses heuristic search but the termination of
the search cannot be intrinsically decided. \code{searchDepth} is needed to
tell the search when to stop.}

\item{method}{The strategy for predicting the fixation. The basic approach is
entropy minimization and can be achieved by adding or removing fixation
point, or by comparing the two.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A \code{fixationSites} object.
}
\description{
After finding the \code{\link{lineagePath}} of a phylogenetic
  tree, \code{fixationSites} uses the result to find those sites that show
  fixation on some, if not all, of the lineages. Parallel evolution is
  relatively common in RNA virus. There is chance that some site be fixed in
  one lineage but does not show fixation because of different sequence
  context.
}
\examples{
data(zikv_tree_reduced)
data(zikv_align_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
fixationSites(lineagePath(tree))
}
\seealso{
\code{\link{as.data.frame.fixationSites}}
}
