% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMutations.R
\name{getSimMatrix}
\alias{getSimMatrix}
\title{Similarity Matrix}
\usage{
getSimMatrix(tree, align)
}
\arguments{
\item{tree}{a \code{phylo} object}

\item{align}{an \code{alignment} object}
}
\description{
To calculate similarity between aligned sequences
}
