% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{getSitePath}
\alias{getSitePath}
\title{Get sitePath}
\usage{
getSitePath(treeAlignMatch, similarity)
}
\arguments{
\item{treeAlignMatch}{a \code{\link{treeAlignMatch}} object}

\item{similarity}{similarity threshold for tree trimming}
}
\value{
path represent by tip names
}
\description{
Under development
}
