% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{groupTips}
\alias{groupTips}
\title{Group Tips}
\usage{
groupTips(tree, align, similarity)
}
\arguments{
\item{tree}{a phylo object}

\item{align}{an alignment object}

\item{similarity}{similarity threshold for tree trimming}
}
\value{
grouping of tips
}
\description{
Group tree tips by their sequence similarity
}
