% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticData.R
\docType{data}
\name{h3n2_align}
\alias{h3n2_align}
\alias{h3n2_align_reduced}
\title{Multiple sequence alignment of H3N2's HA protein}
\format{
an \code{alignment} object

an \code{alignment} object
}
\usage{
data(h3n2_align)

data(h3n2_align_reduced)
}
\description{
The raw protein sequences were downloaded from NCBI database and
  aligned using MAFFT (\url{https://mafft.cbrc.jp/alignment/software/}).

\code{h3n2_align_reduced} is a truncated version of
  \code{h3n2_align}
}
\keyword{datasets}
