% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{h3n2_align_reduced}
\alias{h3n2_align_reduced}
\title{Truncated data for runnable example}
\format{a \code{alignment} object}
\usage{
data(h3n2_align_reduced)
}
\description{
This is a truncated version of \code{\link{h3n2_align}}
}
\keyword{datasets}
