% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{h3n2_tree}
\alias{h3n2_tree}
\title{Phylogenetic tree of H3N2's HA protein}
\format{
a \code{phylo} object
}
\usage{
data(h3n2_tree)
}
\description{
Tree was built from \code{\link{h3n2_align}} using RAxML with
  default settings.
}
\keyword{datasets}
