% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{mutations2Nodes}
\alias{mutations2Nodes}
\title{Format conversion}
\usage{
mutations2Nodes(mutations)
}
\arguments{
\item{mutations}{the return of \code{\link{ancestralMutations}} function}
}
\value{
site and corresponding nodes
}
\description{
Reformat the return from \code{\link{ancestralMutations}}
}
