% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualAdds.R
\name{nodePlot}
\alias{nodePlot}
\title{Internal node plot}
\usage{
nodePlot(mutations)
}
\arguments{
\item{mutations}{the return of \code{\link{phyloMutations}} function}
}
\description{
plot tree with its internal numbering
}
