% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelSites.R
\name{parallelSites}
\alias{parallelSites}
\alias{parallelSites.sitesMinEntropy}
\title{Mutation in multiple lineages}
\usage{
\method{parallelSites}{sitesMinEntropy}(x, minSNP, ...)
}
\arguments{
\item{x}{A \code{\link{sitesMinEntropy}} object.}

\item{minSNP}{Minimum number of tips to have mutations on a lineage.}
}
\value{
A \code{sitesMinEntropy} object
}
\description{
A site may have mutated on parallel lineages.
}
\examples{
data(zikv_tree_reduced)
data(zikv_align_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
paths <- lineagePath(tree)
x <- sitesMinEntropy(paths)
parallelSites(x)
}
