% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.fixationSites}
\alias{plot.fixationSites}
\title{Color the tree by a single site}
\usage{
\method{plot}{fixationSites}(x, y, ...)
}
\arguments{
\item{x}{A \code{fixationSites} object from \code{\link{fixationSites}}}

\item{y}{One of the mutations in the \code{fixationSites} object.
It should be from the \code{\link{names}} of the object.
Or an integer indicating a site could be provide. The numbering
is consistent with the reference defined at 
\code{\link{fixationSites}}.}

\item{...}{Arguments in \code{plot.phylo} functions.}
}
\value{
The function only makes plot and returns no value
(It behaviors like the generic \code{\link{plot}} function).
}
\description{
Visualize \code{fixationSites} object. It will plot the tree
and color the ancestral tips in red, descendant tips in blue and
excluded tips in grey.But the plot will color the tree according to
the amino acid instead if \code{site} argument is provided.
}
\examples{
data("zikv_tree")
data("zikv_align")
tree <- addMSA(zikv_tree, seqs = zikv_align)
fixations <- fixationSites(sitePath(tree, 0.996))
plot(fixations, names(fixations)[[1]])
}
