% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.lineagePath}
\alias{plot.lineagePath}
\title{Visualize phylogenetic lineages}
\usage{
\method{plot}{lineagePath}(x, y = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link{lineagePath}} object}

\item{y}{Whether plot the nodes from the \code{extendedSearch} in
\code{\link{fixationSites}}}

\item{...}{Arguments in \code{plot.phylo} functions.}
}
\value{
The function only makes plot and returns no value
(It behaviors like the generic \code{\link{plot}} function).
}
\description{
Visualize \code{\link{lineagePath}} object. A tree diagram will be plotted
and paths are black solid line while the trimmed nodes and tips will use
grey dashed line.
}
\examples{
data('zikv_tree')
data('zikv_align')
tree <- addMSA(zikv_tree, alignment = zikv_align)
plot(lineagePath(tree, 0.996))
}
