% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualAdds.R
\name{plot.phyloMutations}
\alias{plot.phyloMutations}
\title{Mutation map}
\usage{
\method{plot}{phyloMutations}(mutations)
}
\arguments{
\item{mutations}{the return of \code{\link{phyloMutations}} function}
}
\description{
plot the linked clades and mutations
}
\examples{
\dontrun{
treeFile <- system.file("test.tree", package = "sitePath")
alignFile <- system.file("test.aligned.fasta", package = "sitePath")
matched <- readTreeAlign(treeFile, "newick", alignFile, "fasta")
mutations <- phyloMutations(matched, 0.95, "AA")
plot(mutations)
}
}
