% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.sitePath}
\alias{plot.sitePath}
\title{Plot the fixation mutation}
\usage{
\method{plot}{sitePath}(x, y = NULL, showTips = FALSE, ...)
}
\arguments{
\item{x}{A \code{sitePath} object}

\item{y}{A \code{sitePath} object can have more than one fixation path.
This is to select which path to plot. The default is NULL which
will plot all the paths.}

\item{showTips}{Whether to plot the tip labels. The default is \code{FALSE}.}

\item{...}{Arguments in \code{plot.phylo} functions and other arguments.}
}
\value{
The function only makes plot and returns no value
(It behaviors like the generic \code{\link{plot}} function).
}
\description{
Visualize the \code{sitePath} object which is the basic unit of the
result of \code{\link{fixationSites}} and \code{\link{multiFixationSites}}.
}
\examples{
data(zikv_align_reduced)
data(zikv_tree_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
paths <- lineagePath(tree)
fixations <- fixationSites(paths)
plot(fixations[[1]])
}
\seealso{
\code{\link{plotSingleSite}}
}
