% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixationSites.R
\name{plot.fixationSites}
\alias{plot.fixationSites}
\title{Visualize fixation sites on tree}
\usage{
\method{plot}{fixationSites}(x, y = FALSE, color = NULL, ...)
}
\arguments{
\item{x}{A \code{\link{fixationSites}} object.}

\item{y}{For a \code{\link{fixationSites}} object, it is whether to plot
recurring fixation mutation only. The default is FALSE.}

\item{color}{The color list for the clusters. The default uses ggplot2's
color scheme.}

\item{...}{Other arguments.}
}
\value{
A phylogenetic tree plot marked with site fixation. The function does
  not behave like generic \code{\link{plot}} function.
}
\description{
Mark fixation sites on the original phylogenetic tree.
}
\examples{
data(zikv_tree_reduced)
data(zikv_align_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
paths <- lineagePath(tree)
fixations <- fixationSites(paths)
plot(fixations)
}
