% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFixationSites.R
\name{plotFixationSites}
\alias{plotFixationSites}
\alias{plotFixationSites.fixationSites}
\alias{plotFixationSites.paraFixSites}
\title{Plot the result of fixation sites}
\usage{
plotFixationSites(x, ...)

\method{plotFixationSites}{fixationSites}(x, site = NULL, ...)

\method{plotFixationSites}{paraFixSites}(x, site = NULL, ...)
}
\arguments{
\item{x}{return from \code{\link{paraFixSites}}}

\item{...}{further arguments passed to or from other methods.}

\item{site}{the number of the site according to
\code{\link{setSiteNumbering}}. If not provided, all sites will be plotted
as labels on the tree}
}
\value{
A \code{ggplot} object.
}
\description{
Visualize the results of \code{\link{paraFixSites}}
}
\examples{
data(zikv_tree_reduced)
data(zikv_align_reduced)
paraFix <- paraFixSites(zikv_tree_reduced, alignment = zikv_align_reduced)
plotFixationSites(paraFix)
}
