% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParallelSites.R
\name{plotParallelSites}
\alias{plotParallelSites}
\alias{plotParallelSites.parallelSites}
\alias{plotParallelSites.paraFixSites}
\title{Plot the result of fixation sites}
\usage{
plotParallelSites(x, ...)

\method{plotParallelSites}{parallelSites}(x, site = NULL, ...)

\method{plotParallelSites}{paraFixSites}(x, site = NULL, ...)
}
\arguments{
\item{x}{return from \code{\link{paraFixSites}}}

\item{...}{further arguments passed to or from other methods.}

\item{site}{the number of the site according to
\code{\link{setSiteNumbering}}}
}
\value{
A \code{ggplot} object.
}
\description{
Visualize the results of \code{\link{paraFixSites}}
}
\examples{
data(zikv_tree)
data(zikv_align)
paraFix <- paraFixSites(zikv_tree, alignment = zikv_align)
plotParallelSites(paraFix)
}
