% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineagePath.R
\name{plot.lineagePath}
\alias{plot.lineagePath}
\title{Visualize phylogenetic lineages}
\usage{
\method{plot}{lineagePath}(x, y = TRUE, showTips = FALSE, ...)
}
\arguments{
\item{x}{Could be a \code{\link{lineagePath}} object or the return from
\code{\link{sneakPeek}}.}

\item{y}{Whether to emphasize the lineage branches by using thicker line.}

\item{showTips}{Whether to plot the tip labels. The default is \code{FALSE}.}

\item{...}{Other arguments.}
}
\value{
A ggplot object to plot the tree and dash-line the non-lineage
  branches. The function does not behave like generic \code{\link{plot}}
  function.
}
\description{
Visualize \code{\link{lineagePath}} object. A tree diagram will
  be plotted and paths are black solid line while the trimmed nodes and tips
  will use grey dashed line.
}
\examples{
data(zikv_tree)
data(zikv_align)
tree <- addMSA(zikv_tree, alignment = zikv_align)
paths <- lineagePath(tree)
plot(paths)
}
