% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitewiseClusters.R
\name{plot.sitewiseClusters}
\alias{plot.sitewiseClusters}
\title{Visualize fixation sites}
\usage{
\method{plot}{sitewiseClusters}(x, y = TRUE, ...)
}
\arguments{
\item{x}{Could be a \code{\link{fixationSites}} object or a \code{sitePath}
object.}

\item{y}{For a \code{\link{sitewiseClusters}} object, it is whether to show
the fixation mutation between clusters. For a \code{sitePath} object, it
can have more than one fixation path. This is to select which path to plot.
The default is \code{NULL} which will plot all the paths.}

\item{...}{Other arguments.}
}
\value{
The function only makes plot and returns no value (It behaviors like
  the generic \code{\link{plot}} function).
}
\description{
Visualize \code{\link{fixationSites}} object. The tips are
  clustered according to the fixation sites. The transition of fixation sites
  will be plotted as a phylogenetic tree. The length of each branch
  represents the number of fixation mutation between two clusters. The name
  of the tree tips indicate the number of sequences in the cluster.
}
\examples{
data(zikv_tree_reduced)
data(zikv_align_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
paths <- lineagePath(tree)
fixations <- fixationSites(paths)
plot(fixations)
}
