% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMutations.R
\name{Pre-assessment}
\alias{Pre-assessment}
\alias{similarityMatrix}
\alias{sneakPeek}
\title{Things can be done before the analysis}
\usage{
similarityMatrix(tree, align)

sneakPeek(tree, align, step = NULL, maxPath = NULL, makePlot = TRUE)
}
\arguments{
\item{tree}{a \code{phylo} object}

\item{align}{an \code{alignment} object}

\item{step}{the similarity window}

\item{maxPath}{maximum number of path to show in the plot}

\item{makePlot}{whether make a dot plot when return}
}
\value{
\code{similarityMatrix} returns a diagonal matrix of similarity between sequences

\code{sneakPeek} return the similarity threhold against number of sitePath
}
\description{
\code{similarityMatrix} calculates similarity between aligned sequences
The similarity matrix can be used in \code{\link{groupTips}} or \code{\link{sitePath}}

\code{sneakPeek} is intended to plot similarity as a threshold
against number of output sitePath. This plot is intended to give user
a feel about how many sitePaths they should expect from the similarity threshold.
The number of sitePath should not be too many or too few. The result excludes
where the number of sitePath is greater than number of tips divided by 20 or
self-defined maxPath. The zero sitePath result will also be excluded
}
