% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{readTreeAlign}
\alias{readTreeAlign}
\title{ReadTreeAlign}
\usage{
readTreeAlign(treeFile, treeFormat = c("nexus", "newick", "beast"), alignFile,
  alignFormat = c("fasta", "clustal", "phylip", "mase", "msf"))
}
\arguments{
\item{treeFile}{Path to the file stored tree string}

\item{treeFormat}{Specify tree format}

\item{alignFile}{Path to the file stored sequence alignment}

\item{alignFormat}{Specify alignment format}
}
\value{
a treeAlignMatch object
}
\description{
Read tree and its matched alignment file
}
