% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read.tree}
\alias{read.alignment}
\alias{as.phylo}
\alias{as.treedata}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ape}{\code{\link[ape]{as.phylo}}, \code{\link[ape]{read.tree}}}

  \item{seqinr}{\code{\link[seqinr]{read.alignment}}}

  \item{tidytree}{\code{\link[tidytree]{as.treedata}}}
}}

