% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticData.R
\docType{data}
\name{sars2_align}
\alias{sars2_align}
\title{Multiple sequence alignment of SARS-CoV-2 genome CDS}
\format{
an \code{alignment} object
}
\usage{
data(sars2_align)
}
\description{
The raw sequences were downloaded from GISAID database
  (\url{https://www.gisaid.org/}) and aligned using MAFFT
  (\url{https://mafft.cbrc.jp/alignment/software/}) with default settings.
}
\keyword{datasets}
