% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticData.R
\docType{data}
\name{sars2_tree}
\alias{sars2_tree}
\title{Phylogenetic tree of SARS-CoV-2 genome CDS}
\format{
a \code{phylo} object
}
\usage{
data(sars2_tree)
}
\description{
Tree was built from \code{\link{sars2_align}} using RAxML
  (\url{http://www.exelixis-lab.org/}) with default settings. The tip
  \code{EPI_ISL_402125} was used as the outgroup to root the tree.
}
\keyword{datasets}
