% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupTips.R
\name{similarityMatrix}
\alias{similarityMatrix}
\title{Similarity between sequences}
\usage{
similarityMatrix(tree)
}
\arguments{
\item{tree}{The return from \code{\link{addMSA}} function.}
}
\value{
A diagonal matrix of similarity between sequences.
}
\description{
Get similarity between aligned sequences with gap ignored.
}
\examples{
data(zikv_tree)
data(zikv_align)
tree <- addMSA(zikv_tree, alignment = zikv_align)
simMatrix <- similarityMatrix(tree)
}
