% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitePath-deprecated.R
\name{sitePath-deprecated}
\alias{sitePath-deprecated}
\alias{multiFixationSites}
\title{Deprecated functions in package \sQuote{sitePath}}
\description{
These functions are provided for compatibility with older
  versions of \sQuote{sitePath} only, and will be defunct at the next
  release.
}
\details{
The following functions are deprecated and will be made defunct; use
  the replacement indicated below: \itemize{ \item{multiFixationSites:
  \code{\link{fixationSites}}} }
}
