% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{sitePath}
\alias{sitePath}
\title{Get sitePath}
\usage{
sitePath(tree, align, similarity, simMatrix = NULL)
}
\arguments{
\item{tree}{a \code{phylo} object}

\item{align}{an \code{alignment} object}

\item{similarity}{similarity threshold for tree trimming}

\item{simMatrix}{a diagonal matrix of similarity between sequences}
}
\value{
path represent by tip names
}
\description{
Under development
}
