% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixationSites.R
\name{SNPsites}
\alias{SNPsites}
\title{Finding sites with variation}
\usage{
SNPsites(tree, minSNP = NULL)
}
\arguments{
\item{tree}{The return from \code{\link{addMSA}} function.}

\item{minSNP}{Minimum number of amino acid variation to be a SNP.}
}
\value{
\code{SNPsite} returns a list of qualified SNP site.
}
\description{
Single nucleotide polymorphism (SNP) in the whole package refers
  to variation of amino acid. \code{SNPsite} will try to find SNP in the
  multiple sequence alignment. A reference sequence and gap character may be
  specified to number the site. This is irrelevant to the intended analysis
  but might be helpful to evaluate the performance of
  \code{\link{fixationSites}}.
}
\examples{
data(zikv_tree_reduced)
data(zikv_align_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
SNPsites(tree)
}
