% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{toNodes.phyloMutations}
\alias{toNodes.phyloMutations}
\title{Format conversion}
\usage{
\method{toNodes}{phyloMutations}(mutations)
}
\arguments{
\item{mutations}{A \code{\link{phyloMutations}} object}
}
\value{
site and corresponding nodes
}
\description{
The function reformats a \code{\link{phyloMutations}} object.
Instead of sorting by linked nodes, the mutations are summarized by sites
}
\seealso{
\code{\link{toTips}}
}
