% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{toTips}
\alias{toTips}
\title{Format conversion}
\usage{
toTips(mutations)
}
\arguments{
\item{mutations}{A \code{\link{phyloMutations}} object}
}
\value{
site and corresponding tips
}
\description{
Reformat a \code{\link{phyloMutations}} object.
The whole tree will be chopped at the branches where mutation sits. 
Descendants derived from nodes on the same tree segment are grouped.
}
\seealso{
\code{\link{toNodes}}
}
