% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixationSites.R
\name{plot.fixationSites}
\alias{plot.fixationSites}
\title{Visualize fixation sites}
\usage{
\method{plot}{fixationSites}(
  x,
  y = TRUE,
  showTips = FALSE,
  recurringOnly = FALSE,
  minEffectiveSize = NULL,
  ...
)
}
\arguments{
\item{x}{Could be a \code{\link{fixationSites}} object or a \code{sitePath}
object.}

\item{y}{For a \code{\link{fixationSites}} object, it is whether to show the
fixation mutation between clusters. For a \code{sitePath} object, it can
have more than one fixation path. This is to select which path to plot. The
default is \code{NULL} which will plot all the paths.}

\item{showTips}{Whether to plot the tip labels. The default is \code{FALSE}.}

\item{recurringOnly}{Whether to plot recurring fixation mutation only. The
default is FALSE.}

\item{minEffectiveSize}{The minimum size for a tip cluster in the plot}
}
\description{
Visualize \code{\link{fixationSites}} object. The tips are
  clustered according to the fixation sites. The transition of fixation sites
  will be plotted as a phylogenetic tree. The length of each branch
  represents the number of fixation mutation between two clusters. The name
  of the tree tips indicate the number of sequences in the cluster.
}
\examples{
fixations <- fixationSites(paths)
plot(fixations)
}
\seealso{
\code{\link{as.phylo.fixationSites}}
}
