% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticData.R
\docType{data}
\name{zikv_align_reduced}
\alias{zikv_align_reduced}
\title{Truncated data for runnable example}
\format{
a \code{alignment} object
}
\usage{
data(zikv_align_reduced)
}
\description{
This is a truncated version of \code{\link{zikv_align}}
}
\keyword{datasets}
