% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{zikv_fixations}
\alias{zikv_fixations}
\title{Fixation sites found in Zika virus polyprotein}
\format{a \code{fixationSites} object}
\usage{
data(zikv_fixations)
}
\description{
The fixation mutation predicted on \code{zikv_paths} by
\code{\link{fixationSites}} with default setting.
}
\keyword{datasets}
