% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{zikv_paths}
\alias{zikv_paths}
\title{Phylogenetic lineages of Zika virus polyprotein}
\format{a \code{sitePath} object}
\usage{
data(zikv_paths)
}
\description{
A list of node path to represent phylogenetic lineages of \code{zikv_tree}.
The minimal pairwise sequence similarity was set to 0.996.
}
\keyword{datasets}
