% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{zikv_tree_reduced}
\alias{zikv_tree_reduced}
\title{Truncated data for runnable example}
\format{a \code{phylo} object}
\usage{
data(zikv_tree_reduced)
}
\description{
This is a truncated version of \code{\link{zikv_tree}}
}
\keyword{datasets}
