% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteNumbering.R
\name{phyMSAmatched}
\alias{phyMSAmatched}
\alias{addMSA}
\title{Prepare data for sitePath analysis}
\usage{
addMSA(tree, msaPath = "", msaFormat = "", alignment = NULL)
}
\arguments{
\item{tree}{A \code{\link{phylo}} object. This commonly can be from tree
parsing function in \code{\link{ape}} or \code{\link{ggtree}}. All the
\code{tip.label} should be found in the sequence alignment.}

\item{msaPath}{The file path to the multiple sequence alignment file}

\item{msaFormat}{The format of the multiple sequence alignment file}

\item{alignment}{An \code{alignment} object. This commonly can be from
sequence parsing function in the \code{\link{seqinr}} package. Sequence
names in the alignment should include all \code{tip.label} in the tree}
}
\value{
Since 1.5.12, the function returns a \code{phyMSAmatched} object to
  avoid S3 methods used on \code{phylo} (better encapsulation).
}
\description{
sitePath requires both tree and sequence alignment to do the
  analysis. \code{addMSA} wraps \code{\link{read.alignment}} function in
  \code{\link{seqinr}} package and helps match names in tree and sequence
  alignment. Either provide the file path to an alignment file and its format
  or an alignment object from the return of \code{\link{read.alignment}}
  function. If both the file path and alignment object are given, the
  function will use the sequence in the alignment file.
}
\examples{
data(zikv_tree)
msaPath <- system.file('extdata', 'ZIKV.fasta', package = 'sitePath')
addMSA(zikv_tree, msaPath = msaPath, msaFormat = 'fasta')
}
