% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{ancestralMutations}
\alias{ancestralMutations}
\title{Ancestral Mutations}
\usage{
ancestralMutations(tree, align, similarity, seqType = c("DNA", "AA"),
  model = NULL, siteMode = c(1, 2))
}
\arguments{
\item{tree}{a phylo object}

\item{align}{an alignment object}

\item{similarity}{similarity threshold for tree trimming}

\item{seqType}{specify sequence type ("DNA", "AA")}

\item{model}{substitution model for reconstruction ancestral sequence}

\item{siteMode}{specify the mutational mode of return}
}
\value{
Linked clades and the mutations
}
\description{
Find the mutations in predicted ancestral sequence and corresponding clades
}
