% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{ancestralMutations}
\alias{ancestralMutations}
\title{Ancestral Mutations}
\usage{
ancestralMutations(treeAlignMatch, similarity, seqType = c("DNA", "AA"),
  model = NULL, siteMode = c(1, 2, 3))
}
\arguments{
\item{treeAlignMatch}{a \code{\link{treeAlignMatch}} object}

\item{similarity}{similarity threshold for tree trimming}

\item{seqType}{specify sequence type ("DNA", "AA")}

\item{model}{substitution model for reconstruction ancestral sequence}

\item{siteMode}{specify the mutational mode of return}
}
\value{
Linked clades and the mutations
}
\description{
Find the mutations in predicted ancestral sequence and corresponding clades
}
\examples{
\dontrun{
treeFile <- system.file("test.tree", package = "sitePath")
alignFile <- system.file("test.aligned.fasta", package = "sitePath")
matched <- readTreeAlign(treeFile, "newick", alignFile, "fasta")
mutations <- ancestralMutations(matched, 0.95, "AA")
}
}
