% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeConversion.R
\name{as.data.frame.fixationSites}
\alias{as.data.frame.fixationSites}
\title{Convert results to Data Frame}
\usage{
\method{as.data.frame}{fixationSites}(x, row.names = NULL, optional = FALSE, tipname = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link{fixationSites}} object.}

\item{row.names}{NULL or a character vector giving the row names for the data
frame. Missing values are not allowed.}

\item{optional}{Unimplemented.}

\item{tipname}{Logical: if the return give fixation mutations betweem tip
groups or tip names within a group. The default is \code{FALSE}.}

\item{...}{Other arguments.}
}
\value{
A \code{\link{data.frame}} object.
}
\description{
Convert return of functions in \code{sitePath} package to a
  \code{\link{data.frame}} so can be better worked with.
}
\examples{
data(zikv_tree_reduced)
data(zikv_align_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
fixations <- fixationSites(lineagePath(tree))
as.data.frame(fixations)
}
