% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTips.R
\name{extractTips}
\alias{extractTips}
\alias{extractTips.fixationSites}
\alias{extractTips.multiFixationSites}
\alias{extractTips.sitePath}
\title{Extract info for fixation of a single site}
\usage{
\method{extractTips}{fixationSites}(x, site, select = 1, ...)

\method{extractTips}{multiFixationSites}(x, site, select = 1, ...)

\method{extractTips}{sitePath}(x, select = 1, ...)
}
\arguments{
\item{x}{A \code{fixationSites} or a \code{sitePath} object.}

\item{site}{A site predicted to experience fixation.}

\item{select}{For a site, there theoretically might be more than one fixation
on different lineages. You may use this argument to extract for a specific
fixation of a site. The default is the first fixation of the site.}

\item{...}{Other arguments}
}
\value{
The function \code{extractTips} returns the name of the tips involved
  in the fixation.
}
\description{
The result of \code{\link{fixationSites}} contains all the
  possible sites with fixation mutation. The function \code{extractTips}
  retrieves the name of the tips involved in the fixation.
}
\examples{
data(zikv_tree_reduced)
data(zikv_align_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
mutations <- fixationSites(lineagePath(tree))
extractTips(mutations, 139)
}
