% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMutations.R
\name{findMutations}
\alias{findMutations}
\alias{findFixed.sitePath}
\title{Find Mutations}
\usage{
\method{findFixed}{sitePath}(paths, tolerance = 0, reference = NULL,
  gapChar = "-")
}
\arguments{
\item{paths}{a \code{sitePath} object}
}
\description{
Under development
}
