% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMutations.R
\name{findMutations}
\alias{findMutations}
\alias{findFixed.sitePath}
\title{Find mutation}
\usage{
\method{findFixed}{sitePath}(paths, reference = NULL, gapChar = "-")
}
\arguments{
\item{paths}{a \code{sitePath} object}

\item{reference}{the name of the reference sequence for numbering}

\item{gapChar}{the character to indicate gap}
}
\value{
sites and tip names
}
\description{
Find the fixation site according to the evolutionary relationship.
}
