% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{getSitePath}
\alias{getSitePath}
\title{Get sitePath}
\usage{
getSitePath(tree, align, similarity)
}
\arguments{
\item{tree}{a \code{phylo} object}

\item{align}{an \code{alignment} object}

\item{similarity}{similarity threshold for tree trimming}
}
\value{
path represent by tip names
}
\description{
Under development
}
