% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{groupTips}
\alias{groupTips}
\title{Group Tips}
\usage{
groupTips(treeAlignMatch, similarity)
}
\arguments{
\item{treeAlignMatch}{a \code{\link{treeAlignMatch}} object}

\item{similarity}{similarity threshold for tree trimming}
}
\value{
grouping of tips
}
\description{
Tree tips are grouped by their sequence similarity and members in a group
are constrained to share a same ancestral node.
Similarity between two tips is derived from their multiple sequence alignment.
The site doesn't count into total length if both are gap.
So similarity is calculated as number of matched divided by revised total length
}
\examples{
\dontrun{
treeFile <- system.file("test.tree", package = "sitePath")
alignFile <- system.file("test.aligned.fasta", package = "sitePath")
matched <- readTreeAlign(treeFile, "newick", alignFile, "fasta", "AA")
grouping <- groupTips(matched, 0.95)
}
}
\seealso{
\code{\link{phyloMutations}}
}
