% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{groupTips}
\alias{groupTips}
\title{Group Tips}
\usage{
groupTips(treeAlignMatch, similarity)
}
\arguments{
\item{treeAlignMatch}{a \code{\link{treeAlignMatch}} object}

\item{similarity}{similarity threshold for tree trimming}
}
\value{
grouping of tips
}
\description{
Group tree tips by their sequence similarity
}
\examples{
\dontrun{
treeFile <- system.file("test.tree", package = "sitePath")
alignFile <- system.file("test.aligned.fasta", package = "sitePath")
matched <- readTreeAlign(treeFile, "nexus", alignFile, "fasta")
grouping <- groupTips(matched, 0.95)
}
}
