% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupTips.R
\name{groupTips}
\alias{groupTips}
\alias{groupTips.phyMSAmatched}
\alias{groupTips.sitesMinEntropy}
\alias{groupTips.fixationSites}
\alias{groupTips.fixationPath}
\title{The grouping of tree tips}
\usage{
\method{groupTips}{phyMSAmatched}(
  tree,
  similarity = NULL,
  simMatrix = NULL,
  forbidTrivial = TRUE,
  tipnames = TRUE,
  ...
)

\method{groupTips}{sitesMinEntropy}(tree, tipnames = TRUE, ...)

\method{groupTips}{fixationSites}(tree, tipnames = TRUE, ...)

\method{groupTips}{fixationPath}(tree, tipnames = TRUE, ...)
}
\arguments{
\item{tree}{The return from \code{\link{addMSA}}, \code{\link{lineagePath}}
or other functions in the \code{sitePath} package.}

\item{similarity}{Similarity threshold for tree trimming in \code{groupTips}.
If not provided, the mean similarity subtract standard deviation of all
sequences will be used.}

\item{simMatrix}{A diagonal matrix of similarities for each pair of
sequences.}

\item{forbidTrivial}{Does not allow trivial trimming.}

\item{tipnames}{If return tips as integer or tip names.}

\item{...}{Other arguments.}
}
\value{
\code{groupTips} returns grouping of tips.
}
\description{
\code{groupTips} uses sequence similarity to group tree tips.
  Members in a group are always constrained to share the same ancestral node.
  Similarity between two tips is derived from their multiple sequence
  alignment. The site will not be counted into total length if both are gap.
  Similarity is calculated as number of matched divided by the corrected
  total length. So far the detection of divergence is based on one simple
  rule: the minimal pairwise similarity. The two branches are decided to be
  divergent if the similarity is lower than the threshold.
}
\examples{
data(zikv_tree)
data(zikv_align)
tree <- addMSA(zikv_tree, alignment = zikv_align)
groupTips(tree)
}
