% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{groupTips}
\alias{groupTips}
\title{Trim Tree}
\usage{
groupTips(tree, align, similarity, simMatrix = NULL, tipnames = TRUE)
}
\arguments{
\item{tree}{a \code{phylo} object}

\item{align}{an \code{alignment} object}

\item{similarity}{similarity threshold for tree trimming}

\item{simMatrix}{a diagonal matrix of similarity between sequences}

\item{tipnames}{if return as tipnames}
}
\value{
grouping of tips
}
\description{
Tree tips are grouped by their sequence similarity and members in a group
are constrained to share a same ancestral node.
Similarity between two tips is derived from their multiple sequence alignment.
The site doesn't count into total length if both are gap.
So similarity is calculated as number of matched divided by revised total length
}
