% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{h3n2_align}
\alias{h3n2_align}
\title{Multiple sequence alignment of H3N2's HA protein}
\format{a \code{alignment} object}
\usage{
data(h3n2_align)
}
\description{
The raw protein sequences were downloaded from NCBI database.
}
\keyword{datasets}
