% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{mutations2graphviz}
\alias{mutations2graphviz}
\title{Mutation map}
\usage{
mutations2graphviz(mutations)
}
\arguments{
\item{mutations}{the return of \code{\link{ancestralMutations}} function}
}
\description{
plot the linked clades and mutations
}
