% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{mutations2tips}
\alias{mutations2tips}
\title{Format conversion}
\usage{
mutations2tips(tree, mutations)
}
\arguments{
\item{tree}{a phylo object}

\item{mutations}{the return of \code{\link{ancestralMutations}} function}
}
\value{
clade and corresponding tips
}
\description{
Reformat the return from \code{\link{ancestralMutations}}
}
