% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{mutations2tips}
\alias{mutations2tips}
\title{Get tip names}
\usage{
mutations2tips(tree, mutations)
}
\arguments{
\item{tree}{a phylo object}

\item{mutations}{the return of \code{ancestralMutations} function}
}
\value{
clade and corresponding tips
}
\description{
Find the tips before and after mutation point
}
