% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{mutations2Tips}
\alias{mutations2Tips}
\title{Format conversion}
\usage{
mutations2Tips(mutations)
}
\arguments{
\item{mutations}{the return of \code{\link{ancestralMutations}} function}
}
\value{
site and corresponding tips
}
\description{
Reformat the return from \code{\link{ancestralMutations}}
}
