% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paraFixSites.R
\name{paraFixSites}
\alias{paraFixSites}
\alias{paraFixSites.lineagePath}
\alias{paraFixSites.sitesMinEntropy}
\title{The fixation sites with mutation on parallel lineage}
\usage{
paraFixSites(x, ...)

\method{paraFixSites}{lineagePath}(
  x,
  category = c("intersect", "union", "parallelOnly", "fixationOnly"),
  minEffectiveSize = NULL,
  searchDepth = 1,
  method = c("compare", "insert", "delete"),
  minSNP = NULL,
  mutMode = c("all", "exact", "pre", "post"),
  ...
)

\method{paraFixSites}{sitesMinEntropy}(
  x,
  category = c("intersect", "union", "parallelOnly", "fixationOnly"),
  minSNP = NULL,
  mutMode = c("all", "exact", "pre", "post"),
  ...
)
}
\arguments{
\item{x}{A \code{lineagePath} object returned from \code{\link{lineagePath}}
function.}

\item{...}{further arguments passed to or from other methods.}

\item{category}{Could be \code{parallelOnly}, \code{fixationOnly},
\code{intersect} or \code{union}.}

\item{minEffectiveSize}{The minimum number of tips in a group.}

\item{searchDepth}{The function uses heuristic search but the termination of
the search cannot be intrinsically decided. \code{searchDepth} is needed to
tell the search when to stop.}

\item{method}{The strategy for predicting the fixation. The basic approach is
entropy minimization and can be achieved by adding or removing fixation
point, or by comparing the two.}
}
\value{
A \code{paraFixSites} object.
}
\description{
The operation between the results of \code{\link{fixationSites}}
  and \code{\link{parallelSites}}.
}
\examples{
data(zikv_tree_reduced)
data(zikv_align_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
paraFixSites(lineagePath(tree))
}
