% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelSites.R
\name{parallelSites}
\alias{parallelSites}
\title{Amino acid in parallel lineages}
\usage{
parallelSites(x, minFreq = NULL, refSeqName = NULL)
}
\arguments{
\item{x}{A \code{\link{lineagePath}} object. And amino acid sequence must be
provided.}

\item{minFreq}{Minimum number of an amino acid to be reported and contribute
to make a parallel site. The default is 1.}
}
\value{
A \code{parallelSites} objects.
}
\description{
A site might be mutated to the same amino acid in multiple
  lineages.
}
\examples{
data(zikv_tree_reduced)
data(zikv_align_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
paths <- lineagePath(tree)
parallelSites(paths)
}
