% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelSites.R
\name{parallelSites}
\alias{parallelSites}
\alias{parallelSites.sitesMinEntropy}
\alias{parallelSites.lineagePath}
\title{Mutation in multiple lineages}
\usage{
\method{parallelSites}{sitesMinEntropy}(
  x,
  minSNP = NULL,
  mutMode = c("all", "exact", "pre", "post"),
  ...
)

\method{parallelSites}{lineagePath}(
  x,
  minSNP = NULL,
  mutMode = c("all", "exact", "pre", "post"),
  ...
)
}
\arguments{
\item{x}{A \code{\link{sitesMinEntropy}} object.}

\item{minSNP}{The minimum number of mutations to be qualified as parallel on
at least two lineages. The default is 1.}

\item{mutMode}{The strategy for finding parallel site. The default is to
consider any mutation regardless of the amino acid/nucleotide before and
after mutation; Or the exact same mutation; Or the mutation having amino
acid/nucleotide before or after mutation.}

\item{...}{Other arguments.}
}
\value{
A \code{sitesMinEntropy} object
}
\description{
A site may have mutated on parallel lineages.
}
\examples{
data(zikv_tree_reduced)
data(zikv_align_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
paths <- lineagePath(tree)
x <- sitesMinEntropy(paths)
parallelSites(x)
}
