% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{phyloMutations}
\alias{phyloMutations}
\title{Ancestral Mutations}
\usage{
phyloMutations(treeAlignMatch, similarity, model = NULL, siteMode = c(1, 2,
  3))
}
\arguments{
\item{treeAlignMatch}{a \code{\link{treeAlignMatch}} object}

\item{similarity}{similarity threshold for tree trimming}

\item{model}{substitution model for reconstruction ancestral sequence}

\item{siteMode}{specify the mutational mode of return}
}
\value{
a phyloMutations object
}
\description{
Sequence of each anchestral node is predict by \code{phangorn}.
The tips are grouping by similarity in the same way as \code{\link{groupTips}} does.
Common ancestral nodes of each group become the new terminals.
The tip stays as terminal if it's the only member in the group.
With new terminals, skeleton of the new tree describes the main evolution path.
Mutation events that happened on the main path are collected and
can be categorized as (coevolve, fixation, )
}
\examples{
\dontrun{
treeFile <- system.file("test.tree", package = "sitePath")
alignFile <- system.file("test.aligned.fasta", package = "sitePath")
matched <- readTreeAlign(treeFile, "newick", alignFile, "fasta", "AA")
mutations <- phyloMutations(matched, 0.95)
}
}
\seealso{
\code{\link{groupTips}}
}
