% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.sitePath}
\alias{plot.sitePath}
\title{Visualize phylogenetic lineages}
\usage{
\method{plot}{sitePath}(x, y = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link{sitePath}} object}

\item{y}{Whether plot the nodes from the \code{extendedSearch} in
\code{\link{fixationSites.sitePath}}}

\item{...}{Arguments in \code{plot.phylo} functions.}
}
\value{
The function makes plot but returns no value
(It behaviors like the generic \code{plot} function).
}
\description{
Visualize \code{\link{sitePath}} object. A tree diagram will be plotted
and paths are black solid line while the trimmed nodes and tips will use
grey dashed line.
}
\examples{
data("zikv_tree")
data("zikv_align")
tree <- addMSA(zikv_tree, seqs = zikv_align)
plot(sitePath(tree, 0.996))
}
