% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixationPath.R, R/fixationSites.R,
%   R/lineagePath.R
\name{plot.fixationPath}
\alias{plot.fixationPath}
\alias{plot.fixationSites}
\alias{plot.lineagePath}
\title{Visualize the results}
\usage{
\method{plot}{fixationPath}(x, y = TRUE, ...)

\method{plot}{fixationSites}(x, y = TRUE, ...)

\method{plot}{lineagePath}(x, y = TRUE, showTips = FALSE, ...)
}
\arguments{
\item{x}{Could be a \code{\link{lineagePath}} object,
\code{\link{fixationSites}} object or \code{\link{fixationPath}} object.}

\item{y}{For \code{\link{lineagePath}} object, it is deprecated and no longer
have effect since 1.5.4. For a \code{\link{fixationSites}} or a
\code{\link{fixationPath}} object, it is whether to show the fixation
mutation between clusters.}

\item{...}{Other arguments. Since 1.5.4, the function uses
\code{\link{ggtree}} as the base function to make plots so the arguments in
\code{plot.phylo} will no longer work.}

\item{showTips}{Whether to plot the tip labels. The default is \code{FALSE}.}
}
\value{
A ggplot object to make the plot. A \code{\link{lineagePath}} object
  will be plotted as a tree diagram will be plotted and paths are black solid
  line while the trimmed nodes and tips will use grey dashed line. A
  \code{\link{fixationSites}} object will be plotted as original phylogenetic
  tree marked with fixation substitutions. A \code{\link{fixationPath}}
  object will be plotted as a \code{phylo} object. The tips are clustered
  according to the fixation sites. The transition of fixation sites will be
  plotted as a phylogenetic tree. The length of each branch represents the
  number of fixation mutation between two clusters. The name of the tree tips
  indicate the number of sequences in the cluster.
}
\description{
The plot function to visualize the return of functions in the
  package. Though the function name \code{plot} is used, the plot functions
  here do not behave like the generic \code{\link{plot}} function. The
  underlying function applies \code{\link{ggplot2}}.
}
\examples{
data(zikv_tree)
data(zikv_align)
tree <- addMSA(zikv_tree, alignment = zikv_align)
paths <- lineagePath(tree)
plot(paths)
fixations <- fixationSites(paths)
plot(fixations)
x <- fixationPath(fixations)
plot(x)
}
