% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotSingleSite}
\alias{plotSingleSite}
\alias{plotSingleSite.lineagePath}
\alias{plotSingleSite.fixationSites}
\alias{plotSingleSite.multiFixationSites}
\title{Color the tree by a single site}
\usage{
\method{plotSingleSite}{lineagePath}(x, site, showPath = FALSE,
  reference = NULL, gapChar = "-", ...)

\method{plotSingleSite}{fixationSites}(x, site, ...)

\method{plotSingleSite}{multiFixationSites}(x, site, ...)
}
\arguments{
\item{x}{A \code{fixationSites} object from \code{\link{fixationSites}} or
the return from \code{\link{addMSA}} function.}

\item{site}{One of the mutations in the \code{fixationSites} object. It should
be from the \code{\link{names}} of the object. Or an integer to
indicate a site could be provide. The numbering is consistent with
the reference defined at \code{\link{fixationSites}}.}

\item{showPath}{If plot the lineage result from lineagePath.}

\item{reference}{Name of reference for site numbering. The name has to be one of the
sequences' name. The default uses the intrinsic alignment numbering.}

\item{gapChar}{The character to indicate gap. The numbering will skip the gapChar
for the reference sequence.}

\item{...}{Arguments in \code{plot.phylo} functions and other arguments.}
}
\value{
The function only makes plot and returns no value
(It behaviors like the generic \code{\link{plot}} function).
}
\description{
For \code{lineagePath}, the tree will be colored according to the amino
acid of the site. The color scheme tries to assign distinguishable
color for each amino acid.

For \code{fixationSites}, it will color the ancestral tips in red,
descendant tips in blue and excluded tips in grey.

For \code{multiFixationSites}, it will color the tips which have
their site fixed. The color will use the same amino acid color
scheme as \code{plotSingleSite.lineagePath}
}
\examples{
data("zikv_tree")
data("zikv_align")
tree <- addMSA(zikv_tree, alignment = zikv_align)
paths <- lineagePath(tree, 0.996)
plotSingleSite(paths, 139)
\dontrun{
fixations <- fixationSites(paths)
plotSingleSite(fixations, 139)
}
\dontrun{
multiFixations <- multiFixationSites(paths)
plotSingleSite(multiFixations, 1542)
}
}
