% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMutations.R
\name{pre-assessment}
\alias{pre-assessment}
\alias{similarityMatrix}
\alias{sneakPeek}
\title{Things can be done before the analysis}
\usage{
similarityMatrix(tree)

sneakPeek(tree, step = NULL, maxPath = NULL, makePlot = TRUE)
}
\arguments{
\item{tree}{The return from \code{\link{addMSA}} function}

\item{step}{the similarity window for calculating and ploting}

\item{maxPath}{maximum number of path to show in the plot. The number of path
in the raw tree can be far greater than trimmed tree. To better
see the impact of chaning threshold on path number. This should be
specified. The default is one 20th of tree tip number.}

\item{makePlot}{whether make a dot plot when return}
}
\value{
\code{similarityMatrix} returns a diagonal matrix of
similarity between sequences

\code{sneakPeek} return the similarity threhold against number of sitePath.
There will be a simple dot plot between threshold and path number if
\code{makePlot} is TRUE.
}
\description{
\code{similarityMatrix} calculates similarity between aligned sequences
The similarity matrix can be used in \code{\link{groupTips}}
or \code{\link{sitePath}}

\code{sneakPeek} is intended to plot similarity as a threshold
against number of output sitePath. This plot is intended to give user
a feel about how many sitePaths they should expect from
the similarity threshold. The number of sitePath should not
be too many or too few. The result excludes where the number of sitePath
is greater than number of tips divided by 20 or self-defined maxPath.
The zero sitePath result will also be excluded
}
\examples{
data("zikv_tree")
data("zikv_align")
tree <- addMSA(zikv_tree, seqs = zikv_align)
simMatrix <- similarityMatrix(tree)
sneakPeek(tree)
}
