% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMutations.R
\name{getSimMatrix}
\alias{getSimMatrix}
\alias{SimilarityPlot}
\alias{sneakPeek}
\title{Similarity Matrix}
\usage{
getSimMatrix(tree, align)

sneakPeek(tree, align, step = NULL, maxPath = NULL)
}
\arguments{
\item{tree}{a \code{phylo} object}

\item{align}{an \code{alignment} object}

\item{maxPath}{maximum number of path to show in the plot}

\item{tree}{a \code{phylo} object}

\item{align}{an \code{alignment} object}
}
\value{
a diagonal matrix of similarity between sequences
}
\description{
To calculate similarity between aligned sequences

This function is intended to plot similarity as a threshold
against number of output sitePath. This plot is intended to give user
a feel of how many sitePaths they should expect from the similarity threshold
}
