% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMutations.R
\name{SimilarityPlot}
\alias{SimilarityPlot}
\alias{sneakPeek}
\title{Pre-assessment}
\usage{
sneakPeek(tree, align, simMatrix = NULL, step = NULL, maxPath = NULL)
}
\description{
Under development.This function is intended to plot similarity as a threshold
against number of output sitePath
}
