% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeAlign.R
\name{readTreeAlign}
\alias{readTreeAlign}
\title{Tree and Alignment}
\usage{
readTreeAlign(treeFile, treeFormat = c("nexus", "newick", "beast"),
  alignFile, alignFormat = c("fasta", "clustal", "interleaved",
  "sequential"))
}
\arguments{
\item{treeFile}{Path to the file stored phylogenetic tree}

\item{treeFormat}{Specify tree format ("nexus", "newick", "beast")}

\item{alignFile}{Path to the file stored sequence alignment}

\item{alignFormat}{Specify alignment format ("fasta", "clustal", "interleaved", "sequential")}
}
\value{
a \code{\link{treeAlignMatch}} object
}
\description{
Read tree and its matched alignment file. 
The return type is the input for other analysis in \code{sitePath} package
}
