% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeAlign.R
\name{treeAlignMatch}
\alias{treeAlignMatch}
\title{Tree and Alignment}
\usage{
treeAlignMatch(tree, align)
}
\arguments{
\item{tree}{an object of \code{phylo}}

\item{align}{an object of \code{phyDat}}
}
\value{
a treeAlignMatch object
}
\description{
Assembly a phylo and a phyDat object into a treeAlignMatch object.
The \code{tip.label} of phylo and \code{names} of phyDat should be matched
one by one
}
