% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemer.R
\name{treemer}
\alias{treemer}
\alias{groupTips}
\alias{sitePath}
\alias{customGroupTips}
\alias{customSitePath}
\title{Topology-dependent tree trimming}
\usage{
groupTips(tree, align, similarity, simMatrix = NULL,
  forbidTrivial = TRUE, tipnames = TRUE)

sitePath(tree, align, similarity, simMatrix = NULL,
  forbidTrivial = TRUE)

customGroupTips(tree, align, pvalue = 0.05, forbidTrivial = TRUE,
  tipnames = TRUE)

customSitePath(tree, align, pvalue = 0.05, forbidTrivial = TRUE)
}
\arguments{
\item{tree}{a \code{phylo} object. This commonly can be from tree paring function 
in \code{ape} or \code{ggtree}
All the \code{tip.label} should be found in the alignment}

\item{align}{an \code{alignment} object. This commonly can be from sequence parsing function
in \code{ape} or \code{seqinr} and many others
Sequence names in the alignment should include all \code{tip.label} in the tree}

\item{similarity}{similarity threshold for tree trimming}

\item{simMatrix}{a diagonal matrix of similarity between sequences}

\item{tipnames}{if return as tipnames}

\item{nontrivial}{does not allow trivial trimming}
}
\value{
grouping of tips

path represent by node number
}
\description{
\code{groupTips} uses sequence similarity to group tree tips. Members in a group
are always constrained to share a same ancestral node.
Similarity between two tips is derived from their multiple sequence alignment.
The site will not be counted into total length if both are gap.
Similarity is calculated as number of matched divided by the corrected total length.
So far the detection of divergence is based on one simple rule: the similarity between
two most distant sequence. The two branches are decided to be divergent if the similarity
is lower than the threshold. (Other more statistical involved approaches 
such as Kolmogorov-Smirnov Tests among pair-wise distance could be introduced in the future)

\code{sitePath} finds the sitePath of a phylogenetic tree providing the corresponding sequence alignment.
This is done by trimming the tree to the ancestor node of tips in each group and then
find the the bifurcated terminals of the trimmed tree. The \code{\link{nodepath}} between root node and
the bifurcated terminals is the sitePath. In order to extend the search of mutational site. The sitePath
will tag some of its trailing nodes. Here nodes up to the ancestor of the tips with
the longest \code{\link{nodepath}} are added.
}
