% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualization}
\alias{visualization}
\alias{plot.lineagePath}
\alias{plot.fixationSites}
\alias{plot.sitePath}
\title{Visualize results}
\usage{
\method{plot}{lineagePath}(x, y = TRUE, showTips = FALSE, ...)

\method{plot}{fixationSites}(
  x,
  y = TRUE,
  showTips = FALSE,
  recurringOnly = FALSE,
  minEffectiveSize = NULL,
  ...
)

\method{plot}{sitePath}(x, y = NULL, showTips = FALSE, ...)
}
\arguments{
\item{x}{Could be a \code{\link{lineagePath}} object, a
\code{\link{fixationSites}} object or a \code{sitePath} object.}

\item{y}{For \code{\link{lineagePath}} object, it is deprecated. For a
\code{\link{fixationSites}} object, it is whether to show the fixation
mutation between clusters. For a \code{sitePath} object, it can have more
than one fixation path. This is to select which path to plot. The default
is \code{NULL} which will plot all the paths.}

\item{showTips}{Whether to plot the tip labels. The default is \code{FALSE}.}

\item{...}{Arguments in \code{plot.phylo} functions.}

\item{recurringOnly}{Whether to plot recurring fixation mutation only. The
default is FALSE.}

\item{minEffectiveSize}{The minimum size for a tip cluster in the plot}
}
\value{
The function only makes plot and returns no value (It behaviors like
  the generic \code{\link{plot}} function).
}
\description{
Visualize \code{\link{lineagePath}} object. A tree diagram will
  be plotted and paths are black solid line while the trimmed nodes and tips
  will use grey dashed line.

Visualize \code{\link{fixationSites}} object. The tips are
  clustered according to the fixation sites. The transition of fixation sites
  will be plotted as a phylogenetic tree. The length of each branch
  represents the number of fixation mutation between two clusters. The name
  of the tree tips indicate the number of sequences in the cluster.

Visualize the \code{sitePath} object which can be extracted by
  using \code{\link{extractSite}} on the return of
  \code{\link{fixationSites}} and \code{\link{multiFixationSites}}.
}
\examples{
data(zikv_tree)
data(zikv_align)
tree <- addMSA(zikv_tree, alignment = zikv_align)
paths <- lineagePath(tree)
plot(paths)
fixations <- fixationSites(paths)
plot(fixations)
sp <- extractSite(fixations, 139)
plot(sp)
}
\seealso{
\code{\link{plotSingleSite}}, \code{\link{extractSite}}
}
