% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{zikv_align}
\alias{zikv_align}
\title{Multiple sequence alignment of Zika virus polyprotein}
\format{a \code{alignment} object}
\usage{
data(zikv_align)
}
\description{
The raw protein sequences were downloaded from ViPR database
(\url{https://www.viprbrc.org/}) and aliged using MAFFT.
with default settings.
}
\keyword{datasets}
