% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{zikv_tree}
\alias{zikv_tree}
\title{Phylogenetic tree of Zika virus polyprotein}
\format{a \code{phylo} object}
\usage{
data(zikv_tree)
}
\description{
Tree was built from \code{\link{zikv_align}} using RAxML
with default settings.The tip “ANK57896” was used as
outgroup to root the tree.
}
\keyword{datasets}
